Fix default menu icons not showing up in WP 3.9.
WP 3.9-beta1 removed support for the old CSS class-based "menu-icon-whatever" icons. Everything uses Dashicons now. As a result, we need to display the "admin-*" dashicons in the selector and hide the old icons. For backwards compatibility, we still keep the old icons and use them in WP < 3.9-beta.

Also moved the generic image icon to the end of the list.

git-svn-id: https://plugins.svn.wordpress.org/admin-menu-editor/trunk@875926 b8457f37-d9ea-0310-8a92-e5e31aec5664