Added a workaround for plugin reactivation failing silently when upgrading the plugin in WP 3.8+.
Apparently WordPress changed how it detects plugins recently. Now the built-in upgrader detects the multisite module *before* the main plugin file. As a result, it tries to re-activate that instead of the actual plugin.

I've changed the name of the multisite module to make it sort after the main file. 

git-svn-id: https://plugins.svn.wordpress.org/admin-menu-editor/trunk@888635 b8457f37-d9ea-0310-8a92-e5e31aec5664