Fixed an "array to string conversion" notice caused by passing array data in the query string.
AME used array_diff_assoc() to compare the query string of each menu item to the current URL. array_diff_assoc compares array values by casting them to string, so it will throw a notice if one of the input arrays contains a value that is also an array (and keys match). Fixed by implementing an arrayDiffAssocRecursive() method.

git-svn-id: https://plugins.svn.wordpress.org/admin-menu-editor/trunk@1038205 b8457f37-d9ea-0310-8a92-e5e31aec5664