Fix: Moving a menu item from a built-in parent menu that includes query parameters (e.g. "Pages") to a different submenu or to the top level no longer changes the menu URL.
Details: 

For example, lets say you have a plugin that adds a new item to the "Pages" menu. The item URL will be something like "edit.php?post_type=page&page=plugin_name".  Then you use AME to move the item to the top level. AME does not recognise "edit.php?post_type=page" as a wp-admin file (because of the query). As a result, it will fall back to using "admin.php" as the base URL and generate this URL for the moved item: "admin.php?page=plugin_name".

This patch fixes that edge case by making AME ignore the query and only check the file name.

git-svn-id: https://plugins.svn.wordpress.org/admin-menu-editor/trunk@1172742 b8457f37-d9ea-0310-8a92-e5e31aec5664