Fix a JavaScript error: "TypeError: invalid 'in' operand a". Affects WP 4.3-beta and above.
Cause: 
The menu-editor.js script incorrectly assumes that the default "colors" setting will always be an object. In practice, it can also be false or null when a menu doesn't have any custom colors. The script tries to iterate over the color list using jQuery.each(). In WP 4.2 and below that works fine even if "colors" is false because the version of jQuery included with WordPress handles that invalid input gracefully. However, the version in WP 4.3 throws an exception instead and breaks the script.


  


git-svn-id: https://plugins.svn.wordpress.org/admin-menu-editor/trunk@1213998 b8457f37-d9ea-0310-8a92-e5e31aec5664